tests to make sure all ColumnsTasks run without errors

+++ tests/util.py
@@ -2,25 +2,36 @@
 Util functions for tests
 '''
 
+from subprocess import check_output
+
+try:
+    check_output('createdb test -E UTF8 -T template0', shell=True)
+except:
+    pass
+
+check_output('psql -c "CREATE EXTENSION IF NOT EXISTS postgis"', shell=True)
+
+
+from tasks.util import shell
 from contextlib import contextmanager
 from tasks.meta import (get_engine, current_session, Base, session_commit,
                         session_rollback)
-from sqlalchemy.orm import sessionmaker
 from luigi.worker import Worker
 from luigi.scheduler import CentralPlannerScheduler
-from Queue import Queue
 
 
 def setup():
+    if Base.metadata.bind.url.database != 'test':
+        raise Exception('Can only run tests on database "test"')
     Base.metadata.drop_all()
     Base.metadata.create_all()
-    #current_session().begin()
 
 
 def teardown():
+    if Base.metadata.bind.url.database != 'test':
+        raise Exception('Can only run tests on database "test"')
     current_session().close()
-    Base.metadata.drop_all()
-    #current_session().begin()
+    #Base.metadata.drop_all()
 
 
 
@@ -28,17 +39,11 @@ def runtask(task):
     '''
     Run deps of tasks then the task, faking session management
     '''
-    scheduler = CentralPlannerScheduler(retry_delay=100, remove_delay=1000,
-                                        worker_disconnect_delay=10)
-    with Worker(scheduler=scheduler, worker_id='X') as worker:
-        worker.add(task)
-        worker.run()
-
-    return worker
-    #q = Queue()
-    #worker = TaskProcess(task, 1, Queue())
-    #worker.run()
-    #return q
+    if task.complete():
+        return
+    for dep in task.deps():
+        runtask(dep)
+    task.run()
 
 
 @contextmanager
